requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Limiter==3.12
WTForms==3.2.1
email-validator==2.2.0
PyMySQL==1.1.1
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
gunicorn==23.0.0
